+++ packages/lib/src/types/types.ts
@@ -31,6 +31,7 @@ import type { AnyType } from "./schemas"
 import { typesMaybe, typesMaybeNull } from "./utility/maybe"
 import { OrTypeInfo, typesOr } from "./utility/or"
 import { RefinementTypeInfo, typesRefinement } from "./utility/refinement"
+import { SnapshotProcessorTypeInfo, typesSnapshotProcessor } from "./utility/snapshotProcessor"
 import { typesUnchecked, UncheckedTypeInfo } from "./utility/unchecked"
 export { getTypeInfo } from "./getTypeInfo"
 export { TypeInfo } from "./TypeChecker"
@@ -52,6 +53,7 @@ export {
   ModelTypeInfo,
   OrTypeInfo,
   TupleTypeInfo,
+  SnapshotProcessorTypeInfo,
 }
 
 export const types = {
@@ -79,6 +81,7 @@ export const types = {
   objectMap: typesObjectMap,
   arraySet: typesArraySet,
   tuple: typesTuple,
+  snapshotProcessor: typesSnapshotProcessor,
 
   mapArray<T extends AnyType>(valueType: T) {
     return typesArray(typesTuple(typesString, valueType))