+++ packages/lib/test/types/typeChecking.test.ts
@@ -10,8 +10,10 @@ import {
   ArrayTypeInfo,
   BooleanTypeInfo,
   customRef,
+  fromSnapshot,
   frozen,
   FrozenTypeInfo,
+  getSnapshot,
   getTypeInfo,
   idProp,
   LiteralTypeInfo,
@@ -56,6 +58,7 @@ import { autoDispose } from "../utils"
 beforeEach(() => {
   setGlobalConfig({
     modelAutoTypeChecking: ModelAutoTypeCheckingMode.AlwaysOff,
+    avoidModelTypeInTypedModelSnapshotsIfPossible: false,
   })
 })
 
@@ -323,7 +326,8 @@ test("tuple - simple types", () => {
 
 test("record - simple types", () => {
   const type = types.record(types.number)
-  assert(_ as TypeToData<typeof type>, _ as { [k: string]: number })
+  type T = TypeToData<typeof type>
+  assert(_ as T, _ as { [k: string]: number })
 
   expectTypeCheckOk(type, {})
   expectTypeCheckOk(type, { x: 5, y: 6 })
@@ -1257,4 +1261,140 @@ test("syntax sugar for primitives in tProp", () => {
     // expectTypeCheckFail(type, ss, ["or"], "string | number | boolean")
   }).toThrow(`snapshot '{}' does not match the following type: string | number | boolean`)
   ss.setOr(5)
+})
+
+describe("snapshotProcessor", () => {
+  test("simple", () => {
+    const numberAsStringType = types.snapshotProcessor(
+      types.number,
+      {
+        fromSnapshot(sn: string) {
+          return Number(sn)
+        },
+        toSnapshot(sn): string {
+          return String(sn)
+        },
+      },
+      "numberAsString"
+    )
+
+    const n = fromSnapshot(numberAsStringType, "2")
+    expect(n).toBe(2)
+    const n2 = getSnapshot(numberAsStringType, n)
+    expect(n2).toBe("2")
+  })
+
+  test("complex", () => {
+    const sumObjType = types.object(() => ({
+      a: types.number,
+      b: types.number,
+    }))
+
+    const complexSP = types.snapshotProcessor(
+      sumObjType,
+      {
+        fromSnapshot(sn: string) {
+          return { a: Number(sn) / 2, b: Number(sn) / 2 }
+        },
+        toSnapshot(sn): string {
+          return String(sn.a + sn.b)
+        },
+      },
+      "complexSP"
+    )
+
+    const n = fromSnapshot(complexSP, "4")
+    expect(n.a).toBe(2)
+    expect(n.b).toBe(2)
+    const n2 = getSnapshot(complexSP, n)
+    expect(n2).toBe("4")
+  })
+
+  test("as a model prop", () => {
+    setGlobalConfig({
+      avoidModelTypeInTypedModelSnapshotsIfPossible: true,
+    })
+
+    const numberAsStringType = types.snapshotProcessor(
+      types.number,
+      {
+        fromSnapshot(sn: string) {
+          return Number(sn)
+        },
+        toSnapshot(sn): string {
+          return String(sn)
+        },
+      },
+      "numberAsString"
+    )
+
+    @model("snapshotProcessor/3/M")
+    class M extends Model({
+      p: tProp(numberAsStringType, 10).withSetter(),
+    }) {}
+
+    const tm = types.model(M)
+
+    const m = new M({ p: 30 })
+    expect(getSnapshot(tm, m)).toStrictEqual({
+      p: "30",
+    })
+    m.setP(20)
+    expect(getSnapshot(tm, m)).toStrictEqual({
+      p: "20",
+    })
+
+    const m2 = fromSnapshot(tm, { p: "30" })
+    expect(getSnapshot(tm, m2)).toStrictEqual({
+      p: "30",
+    })
+    m2.setP(20)
+    expect(getSnapshot(tm, m2)).toStrictEqual({
+      p: "20",
+    })
+  })
+
+  test("over a model", () => {
+    setGlobalConfig({
+      avoidModelTypeInTypedModelSnapshotsIfPossible: true,
+    })
+
+    @model("snapshotProcessor/4/M")
+    class M extends Model({
+      p: tProp(types.number).withSetter(),
+    }) {}
+
+    const tm = types.model(M)
+
+    const modelProcessor = types.snapshotProcessor(
+      tm,
+      {
+        fromSnapshot(sn: { p2: string }) {
+          return { p: Number(sn.p2) }
+        },
+        toSnapshot(sn): { p2: string } {
+          return { p2: String(sn.p) }
+        },
+      },
+      "numberAsString"
+    )
+
+    const m = new M({ p: 30 })
+    expect(getSnapshot(modelProcessor, m)).toStrictEqual({
+      p2: "30",
+    })
+    m.setP(20)
+    expect(getSnapshot(modelProcessor, m)).toStrictEqual({
+      p2: "20",
+    })
+
+    const m2 = fromSnapshot(modelProcessor, { p2: "30" })
+    expect(getSnapshot(modelProcessor, m2)).toStrictEqual({
+      p2: "30",
+    })
+    m2.setP(20)
+    expect(getSnapshot(modelProcessor, m2)).toStrictEqual({
+      p2: "20",
+    })
+  })
 })