+++ CHANGELOG.md
@@ -5,8 +5,11 @@
 - Simplified some typings to make them faster.
 
 - When `tProp` is used instead of `prop` then `$modelType` won't be needed in input model snapshots. As a side effect, `$modelType` is now marked as optional in types. Note that for this to work you will also need to use the `fromSnapshot` overload that takes a type as first argument. For compatibility purposes, `getSnapshot` will still return `$modelType` for typed models, but if you want to avoid those too you can do so this way:
+
   1. `setGlobalConfig({ avoidModelTypeInTypedModelSnapshotsIfPossible: true })`
   2. Use the new `getSnapshot` overload that takes a type as first argument.
+
+- Added `types.snapshotProcessor`. Same rules as with typed models above for it to work.
 
 ## 0.63.1